Fix LIMITED-LIMITED HMI level conflict resolution
There was noticed an issue with the conflict resolution
when 1st app has been resumed to FULL and the 2nd app
is being resumed to LIMITED. In that case state controller
wrongly resolves HMI level of 1st application to LIMITED
however it is still can be in FULL.
The lofic of conflict resolution has been updated to
prevent this. Now, state controller will additionally
verify HMI level of resumed application vs conflicting
application, so if the 2nd app was resumed NOT to FULL
then 1st app will keep its current HMI level as this is
not considered to be a conflict.