Fix dangerous work with app commands (#3375)
* SDLCORE-665: Fix dangerous work with app commands

SDLCORE-665

There was related to a thread safety violation
while working with application internal commands
array. This array is a thread safe itself as it is
protected with mutexes, however FindCommand()
function is returing a raw pointer to an internal
array element so the external component is able to
get access to this element at any time even without
locking the mutex.
Core crash which was found in DeleteCommand request
instance is happening by exactly the same reason -
DeleteCommand is using a raw pointer to find the command,
however this command might be destroyed from another
thread at this point of time. As a result, working
thread can access to a destroyed object by pointer.
As a quick solution, each FindCommand() was followed
by a command accessor which prevents unexpected command
destruction. Working thread will use the temporary
copy of command, but not command itself.
However, the better solution would be to change a
design and to avoid using of the raw pointers.

* Update an Application class methods to return by value instead of raw pointer

Updates the next methods:
FindCommand
FindSubMenu
FindChoiceSet

Updates related logic and Unit tests

* fixup! Update an Application class methods to return by value instead of raw pointer

* fixup! Update an Application class methods to return by value instead of raw pointer

* Review fix: log that submenu already exist instead of not found

* fixup! Update an Application class methods to return by value instead of raw pointer

Co-authored-by: ZhdanovP <pzhdanov@luxoft.com>
Co-authored-by: Igor Gapchuk <igapchuck@luxoft.com>