Fix mutex deadlock on VR.ChangeLanguage received (#3370)
* Fix mutex deadlock on VR.ChangeLanguage received

SYNC-10345

There was a problem that on receiving VR.ChangeLanguage SDL is holding application lock via accessor and tries to send some notifications from the same thread.
For sending notifications SDL should check policy permissions for that notification and hance, to acquire policy lock.
At the same time, PTU could happen in the separate thread and during PTU policy lock could be acquired. Also, during PTU SDL will try to acquire applications lock from this thread, however it is acquired by another thread which executing VR.ChangeLanguage request.
As a result we receive thread deadlock. To avoid this deadlock, accessor in VR.ChangeLanguage was scoped in the same manner as for UI and TTS parts.

* Copy apps without redundant scoping data accessor

Co-authored-by: Elvis Kuliiev <ekuliiev@luxoft.com>
Co-authored-by: Aleksandr Kutsan <AKutsan@luxoft.com>