Wait for UsbHandler thread end in UsbHandler dtor
The Thread dedicated for handling libusb events should exit
successfully on it's own.

This guarantees that `libusb_close()` being called on all
devices, as well as `libusb_exit()` call is properly sequenced.

Thread exits in recommended by libusb doc way:
  1. set exit flag
  2. deregister hotplug callbacks, which will wake up
     `libusb_handle_events()` once again.

But this commit changes the way the Thread being joined. It
ensures that `join()` is called with `kNoStop` flag, which
prevents force stop using `pthread_cancel()`.