Fix synchronization in WaypointsPendingResumptionHandler
There was a problem with duplicate subscriptions to shared data
due to time gap between check for already subscribed apps in
SDLRPCPlugin and further processing of corresponding requests in
WayPointsPendingResumptionHandler. That's why this check was moved
to WayPointsPendingResumptionHandler.
Also refactoring of some methods WayPointsPendingResumptionHandler
class was done to make their logic clearer and easier to understand.