Fix a/v streaming retry sequence issues (#3552)
There were noticed two issues related to a/v streaming:
1. Due to object self destruction after TerminateRequest() call
any attempt to access member field of that object may cause an
undefined behavior - it might be a core crash or corrupted value
sometimes. In this particular case, SDL tries to access `message_`
field through `application_id()` function after object destruction.
As a result, sometimes SDL crashes and sometimes this function just
returns 0. Because of that, SDL was not able to find application by
zero id and broke the retry sequence. This causes random failures
of some ATF scripts. To avoid that issue, all retry logic was
extracted into the separate function and `TerminateRequest` was
moved after that function. This will guarantee that there is no
attempts to access object fields after its destruction.
2. There was noticed that SDL makes one redundant retry attempt.
That was because of late retry value increment. To fix that issue,
increment has been placed before retry amount check.