Fix wrong behavior of RCOnRemoteControlSettingsNotification
Looks like `kAllowed` parameter was swithing on/off the RC functionality
even if it was not present and previous value should be taken.
Processing of `kAllowed` and `kAccessMode` was split on two separate
and independent functions make it really isolated.
Processing of both parameters was aligned, so now core does not perform
extra actions if one of parameter is missing.